Original Commit

+++ NS(Attributed)String+Geometrics/NS(Attributed)String+Geometrics.h
@@ -123,7 +123,7 @@
 
 #import <Cocoa/Cocoa.h>
 
-extern int gNSStringGeometricsTypesetterBehavior ;
+extern NSInteger gNSStringGeometricsTypesetterBehavior ;
 
 @interface NSAttributedString (Geometrics)
