+++ NSString+VarArgs.h
@@ -2,13 +2,13 @@
 
 @interface NSString (VarArgs)
 
-- (int)countOccurrencesOfSubstring:(NSString*)substring ;
+- (NSInteger)countOccurrencesOfSubstring:(NSString*)substring ;
 
 /*
  Returns the count of % characters in the receiver, or the
  index of the highest placeholder + 1, (1-10), whichever is larger.
  */
-- (int)countMaxPlaceholders ;
+- (NSInteger)countMaxPlaceholders ;
 
 + (NSString *)replacePlaceholdersInString:(NSString*)s
 								 argPtr_p:(va_list*)argPtr_p ;
@@ -17,7 +17,7 @@
 /*
  returns a string representation of the integer.  Examples: "1" "42", "-579".
  */
-+ (NSString*)stringWithInt:(int)i ;
++ (NSString*)stringWithInt:(NSInteger)i ;
 
 @end
