+++ NSTableView+Scrolling.h
@@ -1,6 +1,5 @@
-
 @interface NSTableView (Scrolling)
 
-- (void)scrollRowToTop:(int)row ;
+- (void)scrollRowToTop:(NSInteger)row ;
 
 @end
